+++ NEWS.md
@@ -1,6 +1,7 @@
 # groomr 0.1.0.9000 (Development Version)
 
 - Added `get_file_sizes()` function to get the length files.
+- Added `replace_caps()` function to replace caps in an atomic vector.
 
 # groomr 0.1.0